refactor: StraightMoveBehavior 추상 클래스를 통한 중복 코드 병합
