+++ src/main/java/janggi/factory/PieceFactory.java
@@ -3,8 +3,8 @@ package janggi.factory;
 import janggi.domain.piece.Piece;
 import janggi.domain.move.Position;
 import janggi.domain.Side;
-import janggi.domain.piece.behavior.Cannon;
-import janggi.domain.piece.behavior.Chariot;
+import janggi.domain.piece.behavior.straightmove.Cannon;
+import janggi.domain.piece.behavior.straightmove.Chariot;
 import janggi.domain.piece.behavior.crossmove.Elephant;
 import janggi.domain.piece.behavior.General;
 import janggi.domain.piece.behavior.Guard;
