+++ src/test/java/janggi/domain/piece/PieceTest.java
@@ -3,7 +3,7 @@ package janggi.domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.Side;
-import janggi.domain.piece.behavior.Cannon;
+import janggi.domain.piece.behavior.straightmove.Cannon;
 import janggi.domain.piece.behavior.Soldier;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
